virtio: Features change to 64 bit in all virtio_dispatch
change all 32 bit features to 64 bit for adapt more features

+++ lib/virtio/virtio.c
@@ -51,7 +51,7 @@ const char *virtio_dev_name(unsigned short devid)
 }
 
 __deprecated void virtio_describe(struct virtio_device *dev, const char *msg,
-				  uint32_t features, struct virtio_feature_desc *desc)
+				  uint64_t features, struct virtio_feature_desc *desc)
 {
 	(void)dev;
 	(void)msg;
